badge component refactoring
